+++ lib/src/db/lib_update_dbs.dart
@@ -373,21 +373,21 @@ Future<void> updateIsoneMraCapacityResults(
   log.info('Updating ISONE MRA results');
   for (var month in months) {
     log.info('Working on month ${month.toIso8601String()}');
-    // var file = archive.getFilename(month);
-    // var url = archive.getUrl(month);
-    // if (download) {
-    //   await baseDownloadUrl(url, file,
-    //       acceptHeader: 'application/json',
-    //       username: dotenv.env['ISONE_WS_USER'],
-    //       password: dotenv.env['ISONE_WS_PASSWORD']);
-    //   log.info('   Downloaded JSON file for ${month.toIso8601String()}');
-    // }
-    // try {
-    //   archive.makeCsvFileForDuckDb(month);
-    //   log.info('   Created CSV files for month ${month.toIso8601String()}');
-    // } catch (e) {
-    //   log.severe(e.toString());
-    // }
+    var file = archive.getFilename(month);
+    var url = archive.getUrl(month);
+    if (download) {
+      await baseDownloadUrl(url, file,
+          acceptHeader: 'application/json',
+          username: dotenv.env['ISONE_WS_USER'],
+          password: dotenv.env['ISONE_WS_PASSWORD']);
+      log.info('   Downloaded JSON file for ${month.toIso8601String()}');
+    }
+    try {
+      archive.makeCsvFileForDuckDb(month);
+      log.info('   Created CSV files for month ${month.toIso8601String()}');
+    } catch (e) {
+      log.severe(e.toString());
+    }
     archive.updateDuckDb(
         months: [month],
         pathDbFile: